Dns decouple
