+++ source/FreeRTOS_UDP_IP.c
@@ -44,6 +44,7 @@
 #include "FreeRTOS_IP.h"
 #include "FreeRTOS_Sockets.h"
 #include "FreeRTOS_IP_Private.h"
+#include "FreeRTOS_IP_Utils.h"
 #include "FreeRTOS_UDP_IP.h"
 #include "FreeRTOS_ARP.h"
 #include "FreeRTOS_DHCP.h"
