+++ test/cbmc/proofs/DNS/DNSlookup/Makefile.json
@@ -10,7 +10,7 @@
   "CBMCFLAGS":
   [
     "--unwind 1",
-    "--unwindset prvProcessDNSCache.0:{CACHE_ENTRIES_UNWIND},strcmp.0:{HOSTNAME_UNWIND},strlen.0:{HOSTNAME_UNWIND},strcpy.0:{HOSTNAME_UNWIND}",
+    "--unwindset FreeRTOS_ProcessDNSCache.0:{CACHE_ENTRIES_UNWIND},strcmp.0:{HOSTNAME_UNWIND},strlen.0:{HOSTNAME_UNWIND},strcpy.0:{HOSTNAME_UNWIND}",
     "--nondet-static"
   ],
   "OBJS":
