+++ test/cbmc/proofs/ParseDNSReply/Makefile.json
@@ -50,7 +50,7 @@
   "CBMCFLAGS":
   [
     "--unwind 1",
-    "--unwindset {PARSELOOP0}:{PARSELOOP0_UNWIND},{PARSELOOP1}:{PARSELOOP1_UNWIND},prvProcessDNSCache.0:5"
+    "--unwindset {PARSELOOP0}:{PARSELOOP0_UNWIND},{PARSELOOP1}:{PARSELOOP1_UNWIND},FreeRTOS_ProcessDNSCache.0:5"
   ],
 
   "OBJS":
