+++ test/unit-test/FreeRTOS_IP_Timers/FreeRTOS_IP_Timers_stubs.c
@@ -67,21 +67,6 @@ UDPPacketHeader_t xDefaultPartUDPPacketHeader =
     }
 };
 
-portINLINE ipDECL_CAST_PTR_FUNC_FOR_TYPE( ICMPHeader_t )
-{
-    return ( ICMPHeader_t * ) pvArgument;
-}
-
-portINLINE ipDECL_CAST_PTR_FUNC_FOR_TYPE( ProtocolPacket_t )
-{
-    return ( ProtocolPacket_t * ) pvArgument;
-}
-
-portINLINE ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPPacket_t )
-{
-    return ( IPPacket_t * ) pvArgument;
-}
-
 void vPortEnterCritical( void )
 {
 }
