+++ test/unit-test/TCPFilePaths.cmake
@@ -7,6 +7,10 @@ set( TCP_SOURCES
      "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_ARP.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_BitConfig.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_DNS.c"
+     "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_DNS_Cache.c"
+     "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_DNS_Callback.c"
+     "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_DNS_Networking.c"
+     "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_DNS_Parser.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_DHCP.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_DHCPv6.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../source/FreeRTOS_ICMP.c"
